scripts/requirements.txt
rembg==2.0.59
pillow==10.3.0
